+++ routes/set/index.tsx
@@ -10,8 +10,18 @@
  * @module
  */
 import { defineRoute } from "$fresh/server.ts";
-import { getAllTournamentSets } from "@/src/startgg.ts";
-import SetCard from "@/components/SetCard.tsx";
+import {
+  getAllTournamentSets,
+  isSetDefined,
+  isSetFinished,
+  isSetPlaying,
+  isSetUnstarted,
+} from "@/src/startgg.ts";
+import {
+  SetDefined,
+  SetFinished,
+  SetUndefined,
+} from "@/components/SetCard.tsx";
 
 /**
  * This route contains a collection of all of the sets in a tournament. The set collection must show
@@ -25,25 +35,29 @@ import SetCard from "@/components/SetCard.tsx";
  */
 export default defineRoute(async (_req, _ctx) => {
   const tournament = (await getAllTournamentSets()).event;
+  const allSets = tournament.phases.map((phase) => phase.sets.nodes).flat();
+  const setPlaying = allSets.filter(isSetPlaying)[0];
   return (
     <div class="px-4 py-8">
       <div class="ml-4">
         <h1 class="text-4xl font-extrabold">{tournament.name}</h1>
-        <code class="text-xs opacity-50">{tournament.id}</code>
       </div>
-      {tournament.phases.map((phase) => (
-        <div class="mt-8">
-          <div class="ml-4 mb-2">
-            <h2 class="text-3xl font-bold">{phase.name}</h2>
-            <code class="text-xs opacity-50">{phase.id}</code>
-          </div>
-          <div class="flex flex-wrap gap-4">
-            {phase.sets.nodes.map((set, index) => (
-              <SetCard set={set} isNext={index === 0} />
-            ))}
-          </div>
-        </div>
-      ))}
+      <div class="mt-16">
+        {setPlaying ? <SetFinished set={setPlaying} /> : null}
+      </div>
+      <div class="flex flex-wrap gap-4 mt-16">
+        {allSets.filter(isSetUnstarted).map((set, index) => (
+          <SetDefined set={set} isNext={index === 0} />
+        ))}
+      </div>
+      <div class="flex flex-wrap gap-4 mt-16">
+        {allSets.filter((set) => !isSetDefined(set)).map((set) => (
+          <SetUndefined set={set} />
+        ))}
+      </div>
+      <div class="flex flex-wrap gap-4 mt-16">
+        {allSets.filter(isSetFinished).map((set) => <SetFinished set={set} />)}
+      </div>
     </div>
   );
 });